memperbaiki user

+++ database/migrations/2025_02_11_003443_add_hp_to_users_table.php
@@ -22,7 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('hp');
+            if (Schema::hasColumn('users', 'hp')) {
+                $table->dropColumn('hp');
+            }
         });
     }
 };
